requirements.txt
streamlit
pandas
openai
openpyxl
opencc-python-reimplemented